moved several calsses to new models and helpers folders
