+++ src/main/java/org/icatproject/ids/helpers/RangeOutputStream.java
@@ -1,4 +1,4 @@
-package org.icatproject.ids;
+package org.icatproject.ids.helpers;
 
 import java.io.FilterOutputStream;
 import java.io.IOException;
