+++ src/main/java/org/icatproject/ids/models/DataFileInfo.java
@@ -1,4 +1,4 @@
-package org.icatproject.ids.v3.models;
+package org.icatproject.ids.models;
 
 import org.icatproject.ids.plugin.DfInfo;
