+++ src/main/java/org/icatproject/ids/thread/DfWriter.java
@@ -9,11 +9,11 @@ import java.util.List;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.icatproject.ids.LockManager.Lock;
+import org.icatproject.ids.models.DataFileInfo;
 import org.icatproject.ids.PropertyHandler;
 import org.icatproject.ids.plugin.ArchiveStorageInterface;
 import org.icatproject.ids.plugin.MainStorageInterface;
 import org.icatproject.ids.v3.FiniteStateMachine.FiniteStateMachine;
-import org.icatproject.ids.v3.models.DataFileInfo;
 
 /**
  * Copies datafiles from main to archive
