+++ src/main/java/org/icatproject/ids/thread/DsRestorer.java
@@ -18,13 +18,13 @@ import org.icatproject.Datafile;
 import org.icatproject.Dataset;
 import org.icatproject.ids.IcatReader;
 import org.icatproject.ids.LockManager.Lock;
+import org.icatproject.ids.helpers.LocationHelper;
+import org.icatproject.ids.models.DataSetInfo;
 import org.icatproject.ids.PropertyHandler;
 import org.icatproject.ids.plugin.ArchiveStorageInterface;
 import org.icatproject.ids.plugin.MainStorageInterface;
 import org.icatproject.ids.plugin.ZipMapperInterface;
 import org.icatproject.ids.v3.FiniteStateMachine.FiniteStateMachine;
-import org.icatproject.ids.v3.helper.LocationHelper;
-import org.icatproject.ids.v3.models.DataSetInfo;
 
 /*
  * Restores datafiles from the slow to the fast storage.
