+++ src/main/java/org/icatproject/ids/v3/DataSelectionForStorageUnitDatafile.java
@@ -8,7 +8,7 @@ import org.icatproject.ids.enums.DeferredOp;
 import org.icatproject.ids.enums.RequestType;
 import org.icatproject.ids.exceptions.InternalException;
 import org.icatproject.ids.exceptions.NotImplementedException;
-import org.icatproject.ids.v3.models.DataInfoBase;
+import org.icatproject.ids.models.DataInfoBase;
 
 public class DataSelectionForStorageUnitDatafile extends DataSelectionV3Base {
