+++ src/main/java/org/icatproject/ids/v3/DataSelectionForStorageUnitDataset.java
@@ -8,8 +8,8 @@ import org.icatproject.ids.enums.DeferredOp;
 import org.icatproject.ids.enums.RequestType;
 import org.icatproject.ids.exceptions.InternalException;
 import org.icatproject.ids.exceptions.NotImplementedException;
-import org.icatproject.ids.v3.models.DataInfoBase;
-import org.icatproject.ids.v3.models.DataSetInfo;
+import org.icatproject.ids.models.DataInfoBase;
+import org.icatproject.ids.models.DataSetInfo;
 
 public class DataSelectionForStorageUnitDataset extends DataSelectionV3Base {
