+++ src/main/java/org/icatproject/ids/v3/PreparedV3.java
@@ -2,7 +2,8 @@ package org.icatproject.ids.v3;
 
 import java.util.Set;
 import java.util.SortedMap;
-import org.icatproject.ids.v3.models.DataInfoBase;
+
+import org.icatproject.ids.models.DataInfoBase;
 
 /* This is a POJO with only package access so don't make data private */
 public class PreparedV3 {
