+++ src/main/java/org/icatproject/ids/v3/handlers/IsReadOnlyHandler.java
@@ -11,9 +11,9 @@ import org.icatproject.ids.exceptions.InsufficientPrivilegesException;
 import org.icatproject.ids.exceptions.InternalException;
 import org.icatproject.ids.exceptions.NotFoundException;
 import org.icatproject.ids.exceptions.NotImplementedException;
+import org.icatproject.ids.helpers.ValueContainer;
 import org.icatproject.ids.v3.RequestHandlerBase;
 import org.icatproject.ids.v3.ServiceProvider;
-import org.icatproject.ids.v3.models.ValueContainer;
 
 public class IsReadOnlyHandler extends RequestHandlerBase {
